Add a mksh configuration
For some time I considered switching from tcsh to a more compatible
shell. Having ruled out most of them I gave mksh a try. It's small,
fast and has nearly all relevant features. After twei weeks with it
I came to the conclusion, that I'm a tcsh guy and to old to switch.
But sometimes it's nice to have a more modern shell available... So,
as compromise, add the mksh configuration to my global dotfiles.