Start an Alacritty instance on C-W-Return.
This is a band aid and hopefully one of the last alteration to my aging
Openbax setup. I'm still using it on my Thinkpad which is an Optimus
device and it will properly die with it... Otherwise I have switched to
Gnome years ago and will likely switch again to Plasma 6 in the near
future.

Nevertheless, I need a better terminal for some tasks than urxvt. But I
can't replace urxvt, because it's rather deeply entangled with the rest
of the desktop.